+++ putt/game.h
@@ -43,9 +43,9 @@ int   game_step(const float[3], float);
 
 void  game_update_view(float);
 
-void  game_set_rot(int);
+void  game_set_rot(float);
 void  game_clr_mag(void);
-void  game_set_mag(int);
+void  game_set_mag(float);
 void  game_set_fly(float);
 
 void  game_ball(int);
